requirements.txt
Package             Version
------------------- -----------
asgiref             3.6.0
async-generator     1.10
attrs               22.2.0
Automat             22.10.0
beautifulsoup4      4.11.1
bs4                 0.0.1
certifi             2022.12.7
cffi                1.15.1
charset-normalizer  2.1.1
constantly          15.1.0
coreapi             2.3.3
coreschema          0.0.4
cryptography        39.0.0
cssselect           1.2.0
Django              4.1.5
djangorestframework 3.14.0
drf-yasg            1.21.4
exceptiongroup      1.1.0
filelock            3.9.0
h11                 0.14.0
hyperlink           21.0.0
idna                3.4
incremental         22.10.0
inflection          0.5.1
itemadapter         0.7.0
itemloaders         1.0.6
itypes              1.2.0
Jinja2              3.1.2
jmespath            1.0.1
lxml                4.9.2
MarkupSafe          2.1.1
outcome             1.2.0
packaging           22.0
parsel              1.7.0
pip                 22.3.1
Protego             0.2.1
psycopg2            2.9.5
pyasn1              0.4.8
pyasn1-modules      0.2.8
pycparser           2.21
PyDispatcher        2.0.6
pyOpenSSL           23.0.0
PySocks             1.7.1
pytz                2022.7
queuelib            1.6.2
requests            2.28.1
requests-file       1.5.1
ruamel.yaml         0.17.21
ruamel.yaml.clib    0.2.7
selenium            4.7.2
service-identity    21.1.0
setuptools          63.4.3
six                 1.16.0
sniffio             1.3.0
sortedcontainers    2.4.0
soupsieve           2.3.2.post1
sqlparse            0.4.3
tldextract          3.4.0
trio                0.22.0
trio-websocket      0.9.2
Twisted             22.10.0
typing_extensions   4.4.0
uritemplate         4.1.1
urllib3             1.26.13
w3lib               2.1.1
wsproto             1.2.0
zope.interface      5.5.2
